requirements.txt
asgiref==3.4.1
astroid==2.6.2
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
behave==1.2.6
behave-django==1.4.0
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
Django==3.2.5
django-cors-headers==3.7.0
django-dotenv==1.4.2
djangorestframework==3.12.4
drf-spectacular==0.17.2
flake8==3.9.2
gitdb==4.0.7
GitPython==3.1.17
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
install==1.3.4
isort==5.9.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==21.0
parse==1.19.0
parse-type==0.5.2
pbr==5.6.0
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.9.3
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.4
pytest-cov==2.12.1
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.7
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.16.0
smmap==4.0.0
soupsieve==2.2.1
sqlparse==0.4.1
stevedore==3.3.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
wrapt==1.12.1
